backend/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile --output-file=requirements.txt pyproject.toml
#
anyio==4.3.0
    # via
    #   httpx
    #   starlette
asn1crypto==1.5.1
    # via decentriq-platform
cbor2==5.6.2
    # via decentriq-platform
certifi==2023.11.17
    # via
    #   decentriq-platform
    #   httpcore
    #   httpx
    #   requests
cffi==1.16.0
    # via cryptography
charset-normalizer==3.3.2
    # via requests
chily==0.8.1
    # via decentriq-platform
click==8.1.7
    # via uvicorn
cryptography==42.0.5
    # via
    #   decentriq-platform
    #   pyopenssl
    #   python-jose
decentriq-dcr-compiler==0.4.0.dev1
    # via decentriq-platform
decentriq-platform==0.24.4
    # via cohort-explorer-backend (pyproject.toml)
defusedxml==0.7.1
    # via odfpy
ecdsa==0.18.0
    # via
    #   decentriq-platform
    #   python-jose
et-xmlfile==1.1.0
    # via openpyxl
exceptiongroup==1.2.0
    # via anyio
fastapi==0.110.0
    # via cohort-explorer-backend (pyproject.toml)
gunicorn==21.2.0
    # via cohort-explorer-backend (pyproject.toml)
h11==0.14.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.4
    # via httpx
httpx==0.27.0
    # via cohort-explorer-backend (pyproject.toml)
idna==3.6
    # via
    #   anyio
    #   httpx
    #   requests
isodate==0.6.1
    # via rdflib
numpy==1.26.4
    # via
    #   pandas
    #   pyarrow
odfpy==1.4.1
    # via pandas
openpyxl==3.1.2
    # via
    #   cohort-explorer-backend (pyproject.toml)
    #   pandas
oxrdflib==0.3.6
    # via cohort-explorer-backend (pyproject.toml)
packaging==23.2
    # via gunicorn
pandas[excel,spss]==2.2.1
    # via
    #   cohort-explorer-backend (pyproject.toml)
    #   pyreadstat
pem==21.2.0
    # via decentriq-platform
protobuf==4.21.9
    # via decentriq-platform
pyarrow==15.0.0
    # via cohort-explorer-backend (pyproject.toml)
pyasn1==0.5.1
    # via
    #   python-jose
    #   rsa
pycparser==2.21
    # via cffi
pydantic==1.10.14
    # via
    #   decentriq-dcr-compiler
    #   fastapi
pyopenssl==24.0.0
    # via decentriq-platform
pyoxigraph==0.3.22
    # via oxrdflib
pyparsing==3.1.1
    # via rdflib
pyreadstat==1.2.6
    # via pandas
python-calamine==0.1.7
    # via pandas
python-dateutil==2.8.2
    # via pandas
python-dotenv==1.0.1
    # via cohort-explorer-backend (pyproject.toml)
python-jose[cryptography]==3.3.0
    # via cohort-explorer-backend (pyproject.toml)
python-multipart==0.0.9
    # via cohort-explorer-backend (pyproject.toml)
pytz==2024.1
    # via pandas
pyxlsb==1.0.10
    # via pandas
pyyaml==6.0.1
    # via cohort-explorer-backend (pyproject.toml)
rdflib==7.0.0
    # via
    #   oxrdflib
    #   sparqlwrapper
requests==2.31.0
    # via decentriq-platform
rsa==4.9
    # via python-jose
sgx-ias-structs==0.1.9
    # via decentriq-platform
six==1.16.0
    # via
    #   ecdsa
    #   isodate
    #   python-dateutil
sniffio==1.3.0
    # via
    #   anyio
    #   httpx
sparqlwrapper==2.0.0
    # via cohort-explorer-backend (pyproject.toml)
sqloxide==0.1.43
    # via decentriq-platform
starlette==0.36.3
    # via fastapi
typing-extensions==4.9.0
    # via
    #   anyio
    #   decentriq-platform
    #   fastapi
    #   pydantic
    #   uvicorn
tzdata==2024.1
    # via pandas
urllib3==2.2.1
    # via requests
uvicorn==0.27.1
    # via cohort-explorer-backend (pyproject.toml)
xlrd==2.0.1
    # via pandas
xlsxwriter==3.2.0
    # via pandas
